# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'API DOCS BOT|POS'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/penjualan/create:
    post:
      summary: 'Penjualan Create'
      operationId: penjualanCreate
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 200
                    message: 'Request Berhasil.'
                  data: {  }
                  request:
                    xxx_id: 1
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 200
                      message:
                        type: string
                        example: 'Request Berhasil.'
                  data:
                    type: object
                    properties: {  }
                  request:
                    type: object
                    properties:
                      xxx_id:
                        type: integer
                        example: 1
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 404
                    message: 'Data Tidak Ditemukan.'
                  data: null
                  request:
                    param_1: 100
                    param_2: example_value
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 404
                      message:
                        type: string
                        example: 'Data Tidak Ditemukan.'
                  data:
                    type: string
                    example: null
                  request:
                    type: object
                    properties:
                      param_1:
                        type: integer
                        example: 100
                      param_2:
                        type: string
                        example: example_value
      tags:
        - Penjualan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                penjualan_pelanggan:
                  type: string
                  description: 'Text biasa.'
                  example: null
                penjualan_tanggal:
                  type: date
                  description: penjualan_tanggal
                  example: '2024-06-14'
                penjualan_total:
                  type: integer
                  description: ''
                  example: 10000
                penjualan_total_bayar:
                  type: integer
                  description: penjualan_total_bayar
                  example: 10000
                penjualan_cara_bayar:
                  type: enum
                  description: 'Contoh [Tunai, Kartu, Kredit, Transfer, skbdn]'
                  example: Tunai
                penjualan_total_kembalian:
                  type: integer
                  description: penjualan_total_bayar
                  example: 10000
                produk_list:
                  type: array
                  description: 'Detail produk'
                  example:
                    - []
                  items:
                    type: object
                    properties:
                      pjual_detail_id:
                        type: integer
                        description: '(Selalu null, flag create)'
                        example: 0
                      pjual_detail_produk_id:
                        type: integer
                        description: 'dari api/produk/list property produk_id'
                        example: 1
                      pjual_detail_qty:
                        type: integer
                        description: ''
                        example: 1
                      pjual_detail_harga:
                        type: integer
                        description: ''
                        example: 1000
                      pjual_detail_diskon:
                        type: integer
                        description: ''
                        example: 1
                      pjual_detail_diskon_rp:
                        type: integer
                        description: ''
                        example: 1000
                      pjual_diskon_subtotal:
                        type: integer
                        description: ''
                        example: 1000
                    required:
                      - pjual_detail_produk_id
                      - pjual_detail_qty
                      - pjual_detail_harga
                      - pjual_detail_diskon
                      - pjual_detail_diskon_rp
                      - pjual_diskon_subtotal
              required:
                - penjualan_pelanggan
                - penjualan_tanggal
                - penjualan_total
                - penjualan_cara_bayar
  /api/penjualan/update:
    put:
      summary: 'Penjualan Update'
      operationId: penjualanUpdate
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 200
                    message: 'Request Berhasil.'
                  data: {  }
                  request:
                    xxx_id: 1
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 200
                      message:
                        type: string
                        example: 'Request Berhasil.'
                  data:
                    type: object
                    properties: {  }
                  request:
                    type: object
                    properties:
                      xxx_id:
                        type: integer
                        example: 1
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 404
                    message: 'Data Tidak Ditemukan.'
                  data: null
                  request:
                    param_1: 100
                    param_2: example_value
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 404
                      message:
                        type: string
                        example: 'Data Tidak Ditemukan.'
                  data:
                    type: string
                    example: null
                  request:
                    type: object
                    properties:
                      param_1:
                        type: integer
                        example: 100
                      param_2:
                        type: string
                        example: example_value
      tags:
        - Penjualan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                penjualan_ida:
                  type: integer
                  description: 'penjualan_ida data dari api/penjualan list.'
                  example: 2
                penjualan_id:
                  type: integer
                  description: 'Text biasa.'
                  example: 1
                penjualan_pelanggan:
                  type: string
                  description: 'Text biasa.'
                  example: null
                penjualan_tanggal:
                  type: date
                  description: penjualan_tanggal
                  example: '2024-06-14'
                penjualan_total:
                  type: integer
                  description: ''
                  example: 10000
                penjualan_total_bayar:
                  type: integer
                  description: penjualan_total_bayar
                  example: 10000
                penjualan_cara_bayar:
                  type: enum
                  description: 'Contoh [Tunai, Kartu, Kredit, Transfer, skbdn]'
                  example: Tunai
                penjualan_total_kembalian:
                  type: integer
                  description: penjualan_total_bayar
                  example: 10000
                produk_list:
                  type: array
                  description: 'Detail produk'
                  example:
                    - []
                  items:
                    type: object
                    properties:
                      pjual_detail_id:
                        type: integer
                        description: '(Selalu null, flag create)'
                        example: 0
                      pjual_detail_produk_id:
                        type: integer
                        description: 'dari api/produk/list property produk_id'
                        example: 1
                      pjual_detail_qty:
                        type: integer
                        description: ''
                        example: 1
                      pjual_detail_harga:
                        type: integer
                        description: ''
                        example: 1000
                      pjual_detail_diskon:
                        type: integer
                        description: ''
                        example: 1
                      pjual_detail_diskon_rp:
                        type: integer
                        description: ''
                        example: 1000
                      pjual_diskon_subtotal:
                        type: integer
                        description: ''
                        example: 1000
                    required:
                      - pjual_detail_produk_id
                      - pjual_detail_qty
                      - pjual_detail_harga
                      - pjual_detail_diskon
                      - pjual_detail_diskon_rp
                      - pjual_diskon_subtotal
              required:
                - penjualan_ida
                - penjualan_id
                - penjualan_pelanggan
                - penjualan_tanggal
                - penjualan_total
                - penjualan_cara_bayar
  '/api/penjualan/delete/{penjualan_id}':
    delete:
      summary: 'Penjualan Delete'
      operationId: penjualanDelete
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 200
                    message: 'Request Berhasil.'
                  data: {  }
                  request:
                    xxx_id: 1
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 200
                      message:
                        type: string
                        example: 'Request Berhasil.'
                  data:
                    type: object
                    properties: {  }
                  request:
                    type: object
                    properties:
                      xxx_id:
                        type: integer
                        example: 1
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status:
                    code: 404
                    message: 'Data Tidak Ditemukan.'
                  data: null
                  request:
                    param_1: 100
                    param_2: example_value
                properties:
                  status:
                    type: object
                    properties:
                      code:
                        type: integer
                        example: 404
                      message:
                        type: string
                        example: 'Data Tidak Ditemukan.'
                  data:
                    type: string
                    example: null
                  request:
                    type: object
                    properties:
                      param_1:
                        type: integer
                        example: 100
                      param_2:
                        type: string
                        example: example_value
      tags:
        - Penjualan
    parameters:
      -
        in: path
        name: penjualan_id
        description: 'penjualan_id data dari api/penjualan list.'
        example: 2
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []